Added detailed stats for today grouped by user.
Refactored the `detailedGitStats` function into two functions.
 - The `detailedGitStats` function has been renamed to
   `_detailedGitStats`.
 - The new `_detailedGitStats` function is almost identical to the
   old function, except it now doesn't output a title and it takes any
   number of arguments that get injected into the `git log` command (via
   the `$@` variable). This increases its flexibility, so that it can be
   used in other places without duplicating code.
 - The `detailedGitStats` function now outputs a title and calls the
   `_detailedGitStats` function with the `$_since` and `$_until`
   variables as parameters (these are passed to the `git log` command as
   described above).

Added the `detailedGitStatsTodayByAuthor` function.
 - This function gets the stats for today grouped by author.
 - It uses the `_detailedGitStats` function to do this by passing in the
   `--since` flag that get passed to `git log`.

Added the `detailedGitStatsTodayByAuthor` to the interactive menu.

Added the `detailedGitStatsTodayByAuthor` to the non-interactive list of
options that can be given as a command line argument.

Added the `detailedGitStatsTodayByAuthor` to the list of possible
arguments.

Tested options 1 and 2 with this repository and output is behaving as
expected.

Tested options 1 and 2 with `_GIT_SINCE` and `_GIT_UNTIL` and output is
behaving as expected.

This should resolve #3.